feature(refactor): rename folder and edit the paths where they are used
Finishes #149579761
