+++ server/controllers/users.js
@@ -2,7 +2,7 @@ import bcrypt from 'bcrypt';
 import localStorage from 'local-storage';
 import jwt from 'jsonwebtoken';
 import jwtoken from '../../jwt';
-import Utils from '../../utils';
+import Utils from '../helper/utils';
 
 require('dotenv').config();
